requirements.txt
appdirs==1.4.3
blessings==1.6
bpython==0.16
curtsies==0.2.11
Django==1.10.6
fancycompleter==0.7
greenlet==0.4.12
numpy==1.12.0
olefile==0.44
ordereddict==1.1
packaging==16.8
pdbpp==0.8.3
Pillow==4.0.0
Pygments==2.2.0
pyparsing==2.2.0
pyrepl==0.8.4
requests==2.13.0
six==1.10.0
wcwidth==0.1.7
wmctrl==0.3
